+++ plugins/svg/src/SvgFeatureRenderer/components/Box.tsx
@@ -4,11 +4,12 @@ import { useTheme } from '@mui/material'
 import { observer } from 'mobx-react'
 
 import Arrow from './Arrow'
-import { getBoxColor, isUTR } from './util'
 
 import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
 import type { Feature, Region } from '@jbrowse/core/util'
 import type { SceneGraph } from '@jbrowse/core/util/layouts'
+import { getBoxColor } from './getBoxColor'
+import { isUTR } from './isUTR'
 
 const utrHeightFraction = 0.65
