+++ plugins/svg/src/SvgFeatureRenderer/components/ProcessedTranscript.tsx
@@ -6,20 +6,33 @@ import { layOutFeature, layOutSubfeatures } from './util'
 
 import type { ExtraGlyphValidator } from './types'
 import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
-import type { Feature, Region } from '@jbrowse/core/util'
+import { getSession, type Feature, type Region } from '@jbrowse/core/util'
 import type { SceneGraph } from '@jbrowse/core/util/layouts'
+import { useFeatureSequence } from '@jbrowse/core/BaseFeatureWidget/SequenceFeatureDetails/hooks'
 
 const ProcessedTranscript = observer(function ProcessedTranscript(props: {
   feature: Feature
   region: Region
   config: AnyConfigurationModel
   featureLayout: SceneGraph
+  displayModel: unknown
   selected?: boolean
   reversed?: boolean
-  [key: string]: unknown
 }) {
-  const { feature, config } = props
+  const { region, displayModel, feature, config } = props
   const subfeatures = getSubparts(feature, config)
+  console.log({ displayModel })
+  const session = getSession(displayModel)
+  const { assemblyName } = region
+  const { sequence, error } = useFeatureSequence({
+    session,
+    assemblyName,
+    feature: feature.toJSON(),
+    upDownBp: 0,
+    forceLoad: true,
+  })
+
+  console.log({ sequence, error })
 
   // we manually compute some subfeatures, so pass these separately
   return <Segments {...props} subfeatures={subfeatures} />
